Organize examples into subfolders
