+++ HtmlForgeX.Examples/Email/ExampleBase64EmbeddingEmail.cs
