+++ HtmlForgeX.Examples/Email/ExampleCorrectedEmailPattern.cs
