+++ HtmlForgeX.Examples/Email/ExampleDecimalMarginEmail.cs
