+++ HtmlForgeX.Examples/Email/ExampleDirectEmailPattern.cs
