+++ HtmlForgeX.Examples/Email/ExampleDirectHeaderFooter.cs
