+++ HtmlForgeX.Examples/Email/ExampleFlexibleHeaderFooter.cs
