+++ HtmlForgeX.Examples/Email/ExampleFlexibleHeaderFooterBuilder.cs
