+++ HtmlForgeX.Examples/Email/ExampleImprovedConsistencyEmail.cs
