+++ HtmlForgeX.Examples/Email/ExampleInvoiceEmail.cs
