+++ HtmlForgeX.Examples/Email/ExampleLayoutConfigurationDemo.cs
