+++ HtmlForgeX.Examples/Email/ExampleLayoutEmailPattern.cs
