+++ HtmlForgeX.Examples/Email/ExampleLinkEncodingEmail.cs
