+++ HtmlForgeX.Examples/Email/ExampleNewsletterEmail.cs
