+++ HtmlForgeX.Examples/Email/ExampleOrderConfirmationEmail.cs
