+++ HtmlForgeX.Examples/Email/ExamplePasswordResetEmail.cs
