+++ HtmlForgeX.Examples/Email/ExampleSimpleConfirmationEmail.cs
