+++ HtmlForgeX.Examples/VisNetwork/AdvancedVisNetwork.cs
@@ -1,4 +1,4 @@
-namespace HtmlForgeX.Examples.Containers;
+namespace HtmlForgeX.Examples.VisNetwork;
 
 internal class AdvancedVisNetwork {
     public static void Demo(bool openInBrowser = false) {
