+++ HtmlForgeX.Examples/VisNetwork/BasicVisNetwork.cs
@@ -1,4 +1,4 @@
-namespace HtmlForgeX.Examples.Containers;
+namespace HtmlForgeX.Examples.VisNetwork;
 
 internal class BasicVisNetwork {
     public static void Demo(bool openInBrowser = false) {
