+++ HtmlForgeX.Examples/VisNetwork/VisNetworkLayoutExamples.cs
@@ -1,4 +1,4 @@
-namespace HtmlForgeX.Examples.Containers;
+namespace HtmlForgeX.Examples.VisNetwork;
 
 internal class VisNetworkLayoutExamples {
     public static void Demo(bool openInBrowser = false) {
