+++ HtmlForgeX.Examples/VisNetwork/VisNetworkPhysicsExamples.cs
@@ -1,4 +1,4 @@
-namespace HtmlForgeX.Examples.Containers;
+namespace HtmlForgeX.Examples.VisNetwork;
 
 internal class VisNetworkPhysicsExamples {
     public static void Demo(bool openInBrowser = false) {
